ASoC: SOF: IPC Abstraction - WIP
Abstract all the IPC specific APIs into a seperate directory structure
so that other IPC mechanisms can be used (detected at FW boot).
