+++ include/sound/sof-ipc-v1/header.h
@@ -111,8 +111,6 @@
 /* Get message component id */
 #define SOF_IPC_MESSAGE_ID(x)			((x) & 0xffff)
 
-/* maximum message size for mailbox Tx/Rx */
-#define SOF_IPC_MSG_MAX_SIZE			384
 
 /*
  * Structure Header - Header for all IPC structures except command structs.
