+++ include/sound/sof-ipc-v1/stream.h
@@ -9,7 +9,7 @@
 #ifndef __INCLUDE_SOUND_SOF_STREAM_H__
 #define __INCLUDE_SOUND_SOF_STREAM_H__
 
-#include <sound/sof/header.h>
+#include <sound/sof-ipc-v1/header.h>
 
 /*
  * Stream configuration.
