+++ sound/soc/sof/trace.c
@@ -12,6 +12,7 @@
 #include <linux/sched/signal.h>
 #include "sof-priv.h"
 #include "ops.h"
+#include "ipc-ops.h"
 
 static size_t sof_trace_avail(struct snd_sof_dev *sdev,
 			      loff_t pos, size_t buffer_size)
@@ -154,73 +155,6 @@ static int trace_debugfs_create(struct snd_sof_dev *sdev)
 	return 0;
 }
 
-int snd_sof_init_trace_ipc(struct snd_sof_dev *sdev)
-{
-	struct sof_ipc_fw_ready *ready = &sdev->fw_ready;
-	struct sof_ipc_fw_version *v = &ready->version;
-	struct sof_ipc_dma_trace_params_ext params;
-	struct sof_ipc_reply ipc_reply;
-	int ret;
-
-	if (!sdev->dtrace_is_supported)
-		return 0;
-
-	if (sdev->dtrace_is_enabled || !sdev->dma_trace_pages)
-		return -EINVAL;
-
-	/* set IPC parameters */
-	params.hdr.cmd = SOF_IPC_GLB_TRACE_MSG;
-	/* PARAMS_EXT is only supported from ABI 3.7.0 onwards */
-	if (v->abi_version >= SOF_ABI_VER(3, 7, 0)) {
-		params.hdr.size = sizeof(struct sof_ipc_dma_trace_params_ext);
-		params.hdr.cmd |= SOF_IPC_TRACE_DMA_PARAMS_EXT;
-		params.timestamp_ns = ktime_get(); /* in nanosecond */
-	} else {
-		params.hdr.size = sizeof(struct sof_ipc_dma_trace_params);
-		params.hdr.cmd |= SOF_IPC_TRACE_DMA_PARAMS;
-	}
-	params.buffer.phy_addr = sdev->dmatp.addr;
-	params.buffer.size = sdev->dmatb.bytes;
-	params.buffer.pages = sdev->dma_trace_pages;
-	params.stream_tag = 0;
-
-	sdev->host_offset = 0;
-	sdev->dtrace_draining = false;
-
-	ret = snd_sof_dma_trace_init(sdev, &params.stream_tag);
-	if (ret < 0) {
-		dev_err(sdev->dev,
-			"error: fail in snd_sof_dma_trace_init %d\n", ret);
-		return ret;
-	}
-	dev_dbg(sdev->dev, "stream_tag: %d\n", params.stream_tag);
-
-	/* send IPC to the DSP */
-	ret = sof_ipc_tx_message(sdev->ipc,
-				 params.hdr.cmd, &params, sizeof(params),
-				 &ipc_reply, sizeof(ipc_reply));
-	if (ret < 0) {
-		dev_err(sdev->dev,
-			"error: can't set params for DMA for trace %d\n", ret);
-		goto trace_release;
-	}
-
-	ret = snd_sof_dma_trace_trigger(sdev, SNDRV_PCM_TRIGGER_START);
-	if (ret < 0) {
-		dev_err(sdev->dev,
-			"error: snd_sof_dma_trace_trigger: start: %d\n", ret);
-		goto trace_release;
-	}
-
-	sdev->dtrace_is_enabled = true;
-
-	return 0;
-
-trace_release:
-	snd_sof_dma_trace_release(sdev);
-	return ret;
-}
-
 int snd_sof_init_trace(struct snd_sof_dev *sdev)
 {
 	int ret;
@@ -279,25 +213,6 @@ page_err:
 	return ret;
 }
 EXPORT_SYMBOL(snd_sof_init_trace);
-
-int snd_sof_trace_update_pos(struct snd_sof_dev *sdev,
-			     struct sof_ipc_dma_trace_posn *posn)
-{
-	if (!sdev->dtrace_is_supported)
-		return 0;
-
-	if (sdev->dtrace_is_enabled && sdev->host_offset != posn->host_offset) {
-		sdev->host_offset = posn->host_offset;
-		wake_up(&sdev->trace_sleep);
-	}
-
-	if (posn->overflow != 0)
-		dev_err(sdev->dev,
-			"error: DSP trace buffer overflow %u bytes. Total messages %d\n",
-			posn->overflow, posn->messages);
-
-	return 0;
-}
 
 /* an error has occurred within the DSP that prevents further trace */
 void snd_sof_trace_notify_for_error(struct snd_sof_dev *sdev)
